fet:restfull, DTO

+++ banco/src/main/java/com/banco/domain/conta/exeption/ContaInativaException.java
@@ -5,7 +5,7 @@ public class ContaInativaException extends RuntimeException {
     private static final String  MENSAGEM_PADRAO = "Conta de número %d não está ativa para receber operações!";
 
 
-    public ContaInativaException(int numeroConta) {
+    public ContaInativaException(Long numeroConta) {
 
         super(String.format(MENSAGEM_PADRAO, numeroConta));
     }
