+++ banco/src/main/java/com/banco/domain/conta/exeption/ContaNaoEncontradaException.java
@@ -3,7 +3,7 @@ package com.banco.domain.conta.exeption;
 public class ContaNaoEncontradaException extends RuntimeException {
     private static final String MENSAGEM_PADRAO = "Conta de número %d não Encontrada!";
 
-    public ContaNaoEncontradaException(int numeroConta) {
+    public ContaNaoEncontradaException(long numeroConta) {
         super(String.format(MENSAGEM_PADRAO, numeroConta));
     }
